Created file for Mitglieder endpoint(s) (#26)
