+++ frontend/src/components/mitglieder/MitgliederTable.tsx
@@ -1,14 +1,7 @@
 import React from "react";
 import BootstrapTable from "react-bootstrap-table-next";
 import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css';
-
-
-export interface Mitglied {
-    first_name: string,
-    last_name: string,
-    party_name: string,
-    wahlkreis_name: string,
-}
+import MitgliederEndpoints, {Mitglied} from "../../rest_client/MitgliederEndpoints";
 
 interface Props {
 
@@ -29,15 +22,13 @@ export class MitgliederTable extends React.Component<Props> {
     }
 
     componentDidMount() {
-        fetch('/api/results/mitglieder')
-            .then(response => response.json())
-            .then(data => {
-                // Sort the data alphabetically by first name
-                data.sort((a: Mitglied, b: Mitglied) => (a.first_name > b.first_name) ? 1 : -1);
-                this.setState({
-                    mitglieder: data,
-                });
+        MitgliederEndpoints.getAll().then(data => {
+            // Sort the data alphabetically by first name
+            data.sort((a, b) => (a.first_name > b.first_name) ? 1 : -1);
+            this.setState({
+                mitglieder: data,
             });
+        });
     }
 
     // TODO: THIS ONLY HAS TO BE DONE ONCE
